% Auto-generated: do not edit by hand
\name{dlLayerGroup}

\alias{dlLayerGroup}

\title{LayerGroup component}

\description{
LayerGroup is a wrapper of LayerGroup in react-leaflet. It takes similar properties to its react-leaflet counterpart.
}

\usage{
dlLayerGroup(children=NULL, attribution=NULL, className=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Attribution}

\item{attribution}{Character. Attribution}

\item{className}{Character. A custom class name to assign to the image. Empty by default.}

\item{id}{Character. The ID used to identify this component in Dash callbacks}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

