% Auto-generated: do not edit by hand
\name{dlMarkerClusterGroup}

\alias{dlMarkerClusterGroup}

\title{MarkerClusterGroup component}

\description{
MarkerClusterGroup is a wrapper of MarkerClusterGroup in react-leaflet. It takes similar properties to its react-leaflet counterpart.
}

\usage{
dlMarkerClusterGroup(children=NULL, className=NULL, id=NULL, options=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component (dynamic)}

\item{className}{Character. A custom class name to assign to the image. Empty by default.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{options}{Named list. Marker cluster group options (a dict). See list of options here
https://github.com/Leaflet/Leaflet.markercluster#all-options}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

