% Auto-generated: do not edit by hand
\name{dlPane}

\alias{dlPane}

\title{Pane component}

\description{
Pane is a wrapper of Pane in react-leaflet. It takes similar properties to its react-leaflet counterpart.
}

\usage{
dlPane(children=NULL, name=NULL, pane=NULL, style=NULL,
className=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{name}{Character. The pane name}

\item{pane}{Character. The leaflet pane of the component}

\item{style}{Named list. The CSS style of the component (dynamic)}

\item{className}{Character. A custom class name to assign to the pane. Empty by default.}

\item{id}{Character. The ID used to identify this component in Dash callbacks}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

