% Auto-generated: do not edit by hand
\name{dlScaleControl}

\alias{dlScaleControl}

\title{ScaleControl component}

\description{
ScaleControl is a wrapper of ScaleControl in react-leaflet. It takes similar properties to its react-leaflet counterpart.
}

\usage{
dlScaleControl(children=NULL, imperial=NULL, metric=NULL,
updateWhenIdle=NULL, maxWidth=NULL, className=NULL, id=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component (dynamic).}

\item{imperial}{Logical. Imperial scale or not.}

\item{metric}{Logical. Metric scale or not.}

\item{updateWhenIdle}{Logical. Update when idle or not.}

\item{maxWidth}{Numeric. Control maxWidth.}

\item{className}{Character. A custom class name to assign to the image. Empty by default.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

